+++ Lib/vc4/Invoke.cpp
@@ -31,9 +31,7 @@ void invoke(int numQPUs, SharedArray<uint32_t> &codeMem, int qpuCodeMemOffset, S
 	// - The final two words are the pointer to the parameters per QPU, and
 	//   the pointer to the kernel program to execute.
 	//
-  int numWords = qpuCodeMemOffset + (2 + params->size() + 1)*numQPUs + 2*numQPUs;
-	//printf("numWords    : %d\n", numWords);
-	//printf("codeMem.size: %d\n", codeMem.size());
+  unsigned numWords = qpuCodeMemOffset + (2 + params->size() + 1)*numQPUs + 2*numQPUs;
 
   assert(numWords < codeMem.size());
 
@@ -59,7 +57,7 @@ void invoke(int numQPUs, SharedArray<uint32_t> &codeMem, int qpuCodeMemOffset, S
     codeMem[offset++] = (uint32_t) qpuCodePtr;
   }
 
-  assertq(offset == numWords, "Check final offset failed");
+  assertq(offset == (int) numWords, "Check final offset failed");
 
   // Launch QPUs
   unsigned result = execute_qpu(mb, numQPUs, (uint32_t) launchMsgsPtr, 1, QPU_TIMEOUT);
